@@ -9,11 +9,11 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.2.9</version>
+        <version>3.5.7</version>
         <relativePath /> <!-- lookup parent from repository -->
     </parent>
 
-    <version>6.0.1-SNAPSHOT</version>
+    <version>6.1.0-SNAPSHOT</version>
     <description>LTI tool which exports sis grades</description>
     <url>https://github.com/indiana-university/lms-canvas-sis-grades-export</url>
     <inceptionYear>2015</inceptionYear>
@@ -56,22 +56,23 @@
         <app.version>${project.version}</app.version>
         <commons-collections4.version>4.4</commons-collections4.version>
         <commons-io.version>2.8.0</commons-io.version>
-        <denodo-hibernate-dialect.version>8.0.20201021</denodo-hibernate-dialect.version>
-        <denodo-jdbc.version>8.0.2</denodo-jdbc.version>
+        <denodo-hibernate-dialect.version>9.0.20240513</denodo-hibernate-dialect.version>
+        <denodo-jdbc.version>9.1.0</denodo-jdbc.version>
         <frontend-maven-plugin.version>1.11.0</frontend-maven-plugin.version>
-        <git-commit-id-maven-plugin.version>9.0.1</git-commit-id-maven-plugin.version>
+        <git-commit-id-maven-plugin.version>9.0.2</git-commit-id-maven-plugin.version>
         <hibernate6-ddl-maven-plugin.version>1.0.2</hibernate6-ddl-maven-plugin.version>
-        <java.version>21</java.version>
-        <jdk.source>21</jdk.source>
-        <jdk.target>21</jdk.target>
-        <lms-embedded-services.version>6.0.2</lms-embedded-services.version>
-        <lms-team-spring-boot-it12>6.5.0</lms-team-spring-boot-it12>
-        <opencsv.version>5.9</opencsv.version>
-        <spring-cloud-starter-parent.version>2023.0.3</spring-cloud-starter-parent.version>
-        <springdoc-openapi-ui.version>2.6.0</springdoc-openapi-ui.version>
+        <java.version>25</java.version>
+        <jdk.source>25</jdk.source>
+        <jdk.target>25</jdk.target>
+        <lms-embedded-services.version>6.3.2</lms-embedded-services.version>
+        <lms-team-spring-boot-it12>6.5.2_2</lms-team-spring-boot-it12>
+        <logcaptor.version>2.12.1</logcaptor.version>
+        <opencsv.version>5.12.0</opencsv.version>
+        <spring-cloud-dependencies.version>2025.0.0</spring-cloud-dependencies.version>
+        <springdoc-openapi-ui.version>2.8.14</springdoc-openapi-ui.version>
 
-        <plugins.compiler.version>3.13.0</plugins.compiler.version>
-        <plugins.license.version>2.4.0</plugins.license.version>
+        <plugins.compiler.version>3.14.1</plugins.compiler.version>
+        <plugins.license.version>2.7.0</plugins.license.version>
         <plugins.release.version>3.1.1</plugins.release.version>
     </properties>
 
@@ -79,8 +80,8 @@
         <dependencies>
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
-                <artifactId>spring-cloud-starter-parent</artifactId>
-                <version>${spring-cloud-starter-parent.version}</version>
+                <artifactId>spring-cloud-dependencies</artifactId>
+                <version>${spring-cloud-dependencies.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
@@ -187,6 +188,12 @@
             <artifactId>spring-boot-properties-migrator</artifactId>
             <scope>runtime</scope>
         </dependency>
+        <dependency>
+            <groupId>io.github.hakky54</groupId>
+            <artifactId>logcaptor</artifactId>
+            <version>${logcaptor.version}</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
@@ -258,6 +265,7 @@
                 <configuration>
                     <source>${java.version}</source>
                     <target>${java.version}</target>
+                    <proc>full</proc>
                 </configuration>
             </plugin>
             <plugin>
